Rename FuzzyInBuffer to FuzzySearch
Ok, this might seem contentious, but... renaming to FuzzySearch removes
any confusion between FuzzyBuffers and FuzzyInBuffer, pretty accurately
describes what it is, i.e. a fuzzy version of Vim's search command, and
allows for a simple mnemonic for a default leader mapping, <leader>fs.

Before this rename I was struggling to differentiate between searching
the buffer list and searching within the current buffer, and coming up
with both decent, short and accurate, names for the commands and nice
mnemonics for association of command name to leader mappings. I'm still
not convinced of this, naming is hard, open to alternative suggestions.